Improved version comparison code
+ fixed small error

The version comparison code now converts each part to an integer instead
of doing the comparison as a string. The original implementation would
be buggy if we ever compared "1.9" to "1.10".

A side-effect of this, though, is that version numbers _must_ be
dot-separated integers. I've been using, for instance, "1.3a" to mean
I'm _working on_ 1.3 but am not ready to call it 1.3 yet.

So now I will adopt this convention:
 * 1.3.1, 1.3.2 etc. will mean that I have _released_ 1.3 and am now
   working towards 1.4.
 * Only 1.3, 1.4, ... get released, not 1.3.1, 1.3.2, ...
 * (In practice, I'll probably only use 1.3.1, but we'll see.)