Added feature: l.Admin.quiet_test_all()
This is not revealed to students. It's for testing: to get a quick check
on all exercises.

l.Admin.quiet_test(205) gives True/False for that exercise without any
printed output. But quiet_test_all() is probably more commonly used, and
it just indicates which ones are not fully passing.